# doc/api.yaml
openapi: 3.0.0
paths:
...
  "/track/{id}":
    get:
      operationId: findOne
      summary: Get single track by id
      description: Gets single track by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad request. trackId is invalid (not uuid)
        "404":
          description: Track was not found
      tags: *a2
    put:
      operationId: update
      summary: Update track information
      description: Update library track information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad request. trackId is invalid (not uuid)
        "404":
          description: Track was not found
      tags: *a2
    delete:
      operationId: remove
      summary: Delete track
      description: Delete track from library
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Deleted successfully
        "400":
          description: Bad request. trackId is invalid (not uuid)
        "404":
          description: Track was not found
      tags: *a2
...
  "/favs/track/{id}":
    post:
      operationId: createTrackFav
      summary: Add track to the favorites
      description: Add track to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: Added successfully
        "400":
          description: Bad. trackId is invalid (not uuid)
        "422":
          description: Track with id doesn't exist
      tags: *a3
    delete:
      operationId: removeTrackFav
      summary: Delete track from favorites
      description: Delete track from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Deleted successfully
        "400":
          description: Bad. trackId is invalid (not uuid)
        "404":
          description: Track was not found
      tags: *a3
  "/favs/album/{id}":
    post:
      operationId: createAlbumFav
      summary: Add album to the favorites
      description: Add album to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: Added successfully
        "400":
          description: Bad. albumId is invalid (not uuid)
        "422":
          description: Album with id doesn't exist
      tags: *a3
    delete:
      operationId: removeAlbumFav
      summary: Delete album from favorites
      description: Delete album from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Deleted successfully
        "400":
          description: Bad. albumId is invalid (not uuid)
        "404":
          description: Album was not found
      tags: *a3
  "/favs/artist/{id}":
    post:
      operationId: createArtistFav
      summary: Add artist to the favorites
      description: Add artist to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: Added successfully
        "400":
          description: Bad request. artistId is invalid (not uuid)
        "422":
          description: Artist with id doesn't exist
      tags: *a3
    delete:
      operationId: removeArtistFav
      summary: Delete artist from favorites
      description: Delete artist from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Deleted successfully
        "400":
          description: Bad request. artistId is invalid (not uuid)
        "404":
          description: Artist was not found
      tags: *a3
  /artist:
    post:
      operationId: create
      summary: Add new artist
      description: Add new artist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtistDto"
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Bad request. body does not contain required fields
      tags: &a4
        - Artist
    get:
      operationId: findAll
      summary: Get all artists
      description: Gets all artists
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
      tags: *a4
  "/artist/{id}":
    get:
      operationId: findOne
      summary: Get single artist by id
      description: Get single artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Bad request. artistId is invalid (not uuid)
        "404":
          description: Artist was not found
      tags: *a4
    put:
      operationId: update
      summary: Update artist information
      description: Update artist information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtistDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Bad request. artistId is invalid (not uuid)
        "404":
          description: Artist was not found
      tags: *a4
    delete:
      operationId: remove
      summary: Delete artist
      description: Delete artist from library
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: ""
        "400":
          description: Bad request. artistId is invalid (not uuid)
        "404":
          description: Artist was not found
      tags: *a4
  /album:
    post:
      operationId: create
      summary: Add new album
      description: Add new album information
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumDto"
      responses:
        "201":
          description: Album is created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Bad request. body does not contain required fields
      tags: &a5
        - Album
    get:
      operationId: findAll
      summary: Get albums list
      description: Gets all library albums list
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"
      tags: *a5
  "/album/{id}":
    get:
      operationId: findOne
      summary: Get single album by id
      description: Gets single album by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Bad request. albumId is invalid (not uuid)
        "404":
          description: Album was not found
      tags: *a5
    put:
      operationId: update
      summary: Update album information
      description: Update library album information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Bad request. albumId is invalid (not uuid)
        "404":
          description: Album was not found
      tags: *a5
    delete:
      operationId: remove
      summary: Delete album
      description: Delete album from library
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Deleted successfully
        "400":
          description: Bad request. albumId is invalid (not uuid)
        "404":
          description: Album was not found
      tags: *a5
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0
  contact: {}
...
servers:
  - url: /api
components:
  schemas:
...
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
          nullable: true
        albumId:
          type: string
          nullable: true
        duration:
          type: number
      required:
        - name
        - duration
    Track:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        artistId:
          type: string
          nullable: true
        albumId:
          type: string
          nullable: true
        duration:
          type: number
      required:
        - id
        - name
        - duration
    Artist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        grammy:
          type: boolean
      required:
        - id
        - name
        - grammy
    Album:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          nullable: true
      required:
        - id
        - name
        - year
...
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          nullable: true
      required:
        - name
        - year